Fix code format.
