+++ fluid/face_detction/pyramidbox.py
@@ -45,7 +45,7 @@ def conv_block(input, groups, filters, ksizes, strides=None, with_pool=True):
 
 
 class PyramidBox(object):
-    def __init__(self, data_shape, is_infer=False, flag_sub_network=False):
+    def __init__(self, data_shape, is_infer=False, sub_network=False):
         self.data_shape = data_shape
         self.min_sizes = [16., 32., 64., 128., 256., 512.]
         self.steps = [4., 8., 16., 32., 64., 128.]
@@ -54,7 +54,7 @@ class PyramidBox(object):
         # the base network is VGG with atrus layers
         self._input()
         self._vgg()
-        if flag_sub_network:
+        if sub_network:
             self._low_level_fpn()
             self._cpm_module()
             self._pyramidbox()
@@ -242,7 +242,10 @@ class PyramidBox(object):
         self.conv5_norm = self._l2_norm_scale(self.conv5)
 
         mbox_locs, mbox_confs, box, box_var = fluid.layers.multi_box_head(
-            inputs=[self.conv3_norm, self.conv4_norm, self.conv5_norm, self.conv6, self.conv7, self.conv8],
+            inputs=[
+                self.conv3_norm, self.conv4_norm, self.conv5_norm, self.conv6,
+                self.conv7, self.conv8
+            ],
             image=self.image,
             num_classes=num_classes,
             # min_ratio=20,
@@ -254,13 +257,13 @@ class PyramidBox(object):
             steps=[4.0, 8.0, 16.0, 32.0, 64.0, 128.0],
             base_size=image_shape[2],
             offset=0.5,
-            flip=False)        
+            flip=False)
 
         # locs, confs, box, box_var = vgg_extra_net(num_classes, image, image_shape)
         # nmsed_out = fluid.layers.detection_output(
-            # locs, confs, box, box_var, nms_threshold=args.nms_threshold)
-        loss = fluid.layers.ssd_loss(mbox_locs, mbox_confs, self.gt_box, self.gt_label, box,
-                                     box_var)
+        # locs, confs, box, box_var, nms_threshold=args.nms_threshold)
+        loss = fluid.layers.ssd_loss(mbox_locs, mbox_confs, self.gt_box,
+                                     self.gt_label, box, box_var)
         loss = fluid.layers.reduce_sum(loss)
 
         return loss
